+++ setup.cfg
@@ -37,7 +37,7 @@ octavia.api.drivers =
 octavia.plugins =
     f5_plugin = octavia_f5.controller.worker.controller_worker:ControllerWorker
 octavia.plugins.notifications =
-    f5_plugin = octavia_f5.controller.worker.controller_worker:ControllerWorkerNotifications
+    f5_plugin = octavia_f5.controller.worker.controller_worker_notifications:ControllerWorkerNotifications
 octavia.network.drivers =
     neutron_client = octavia_f5.network.drivers.neutron.neutron_client:NeutronClient
     network_noop_driver_f5 = octavia_f5.network.drivers.noop_driver_f5.driver:NoopNetworkDriverF5
